docker/docker-tensorflow/requirements.txt
scipy==0.19.0
numpy==1.12.1
atari-py==0.0.21 # gym[atari]==0.8.2 should use 0.0.21 https://github.com/openai/gym#what-s-new
gym[atari]==0.8.2
tensorflow==1.2.1
pillow==4.1.1